+++ tracetypes/org.eclipse.tracecompass.incubator.ros2.core/src/org/eclipse/tracecompass/incubator/internal/ros2/core/model/messages/Ros2TimerCallbackInstance.java
@@ -27,7 +27,7 @@ import org.eclipse.tracecompass.incubator.internal.ros2.core.model.objects.Ros2O
 public class Ros2TimerCallbackInstance extends Ros2Instance<@NonNull Ros2ObjectHandle> {
 
     /** Custom type value ID for this object */
-    public static final byte CUSTOM_TYPE_ID = 69;
+    public static final short CUSTOM_TYPE_ID = 69;
     /** CustomStateValueFactory for this object */
     @SuppressWarnings("restriction")
     public static final @NonNull CustomStateValueFactory ROS2_TIMER_CALLBACK_INSTANCE_VALUE_FACTORY = Ros2TimerCallbackInstance::read;
@@ -90,7 +90,7 @@ public class Ros2TimerCallbackInstance extends Ros2Instance<@NonNull Ros2ObjectH
     }
 
     @Override
-    protected @NonNull Byte getCustomTypeId() {
+    protected @NonNull Short getCustomTypeId() {
         return CUSTOM_TYPE_ID;
     }
